requirements.txt
requests==2.31.0
geojson==3.1.0
python-dateutil==2.8.2
pytz==2023.3
python-dotenv
beautifulsoup4
